Tiny fix on toc issues accross content groups. Needs more work.
Here's the problem: we load one single markdown parser.
This causes issues across content groups for our table of contents.

Because we only have one parser instance, we also only have 1 callback to handle
our toc content. Since the parser is called from the content/page.js, the callback has
that context.

When we move to the next content group, that conntext still points to something
that is now not very useful. Or at least that's how I understand it ;)

The best solution would be to spawn a parser per content group.
This would also allow one to configure your markdown parser per content group,
which king of makes sense. As in, you may want to load different markdown-it plugins
for one type of content vs another.

I'm absolutely not certain how to do this, but I'll keep working on it.