Normalizes slugs to avoid problems with accents in content filenames.
This fixes an issue reported by Sandro where accents in filenames cause
a page not found error. Here's what was going on:

Given a content file named `2018-04-05-test-germän-stüff.md` in section
`en/blog` to be served as `/blog/:slug`, this is what happened:

The slug generated from the filename is `test-germ%C3%A4n-st%C3%BCff`
which is the urlencode version of `test-germän-stüff`

The permalink genrated is the same, but to fetch this from the
content-api, you would need to re-encode it. Specifically, this will not work:

```
/content-api/blog/en/blog/test-germ%C3%A4n-st%C3%BCff
```
Because it will get decoded to `/content-api/blog/en/blog/test-germän-stüff` which does not exist.

Instead, you would have to double encode it, and fetch this:

```
/content-api/en/blog/blog/test-germ%25C3%25A4n-st%25C3%25BCff
```

That does work.

There's two options to deal with this:

**Urlencode the permalink before fetching it from the content-api**

This has the benefit of being able to use `test-germän-stüff` in the
browser URL bar (if that's important for you).

**Normalize the filename**

Just get rid of characters that can wreak havoc,
and normalize the slug to test-german-stuff.

This patch implements the latter approach.

Arguably, this means that people who want fancy accents in the URL bar
are shit out of luck. If that happens to be an issue down the line, I
guess we could try to make this an option that's configurable.